backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.2.0
sqlalchemy==2.0.37
alembic==1.14.1
psycopg[binary]==3.2.3
pandas==2.2.3
openpyxl==3.1.5
xlsxwriter==3.2.0
gunicorn==23.0.0
structlog==24.4.0
